fix: quizzes 페이지 suspense로 CSR으로 변경
- useSearchParams 으로 인해, SSR으로 진행 시, build error 발생하여 CSR로
  변경
